Extract getBitmap() to a new class to avoid duplication
